requirements.txt
flask
appwrite
gunicorn
pandas
python-dotenv